+++ services/payments/authorize.go
@@ -1,0 +1,106 @@
+package payments
+
+import (
+	"context"
+	"encoding/hex"
+	"errors"
+	"fmt"
+
+	"github.com/brave-intl/bat-go/libs/httpsignature"
+	. "github.com/brave-intl/bat-go/libs/payments"
+)
+
+// validAuthorizers is the list of payment authorizers, mapping to individuals in payments-ops.
+var validAuthorizers = map[string]bool{
+	// test/private.pem
+	"a5700b95f77fa0fc078cd923ad5075a100d6b995ecc86e49919a0f6ee45ee983": true,
+	// test/private2.pem
+	"732afdb29da6d5ab8481b247d9b2724d79c3652dddc64eb5ad251a2679e6210d": true,
+}
+
+// LookupVerifier implements keystore for httpsignature.
+func (s *Service) LookupVerifier(ctx context.Context, keyID string) (context.Context, *httpsignature.Verifier, error) {
+	// keyID is the public key, we need to see if this exists in our verifier map
+	if allowed, exists := validAuthorizers[keyID]; !exists || !allowed {
+		return nil, nil, &ErrInvalidAuthorizer{}
+	}
+
+	var (
+		publicKey httpsignature.Ed25519PubKey
+		err       error
+	)
+
+	publicKey, err = hex.DecodeString(keyID)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to decode verifier public key: %w", err)
+	}
+
+	verifier := httpsignature.Verifier(publicKey)
+	return ctx, &verifier, nil
+}
+
+// AuthorizeTransaction - Add an Authorization for the Transaction and attempt to Drive
+// the Transaction forward. NOTE: This function assumes that the http signature has been
+// verified before running. This is achieved in the SubmitHandler middleware.
+func (s *Service) AuthorizeTransaction(
+	ctx context.Context,
+	keyID string,
+	transaction AuthenticatedPaymentState,
+) error {
+	fetchedTxn, idempotencyKey, err := s.GetTransactionFromDocumentID(ctx, transaction.DocumentID)
+	if err != nil {
+		return fmt.Errorf(
+			"failed to get transaction with idempotencyKey %s by document ID %s: %w",
+			idempotencyKey,
+			transaction.DocumentID,
+			err,
+		)
+	}
+	auth := PaymentAuthorization{
+		KeyID:      keyID,
+		DocumentID: transaction.DocumentID,
+	}
+	keyHasNotYetSigned := true
+	for _, authorization := range fetchedTxn.Authorizations {
+		if authorization.KeyID == auth.KeyID {
+			keyHasNotYetSigned = false
+		}
+	}
+	if !keyHasNotYetSigned {
+		return fmt.Errorf("key %s has already signed document %s", auth.KeyID, fetchedTxn.DocumentID)
+	}
+	fetchedTxn.Authorizations = append(fetchedTxn.Authorizations, auth)
+	authenticatedState, err := writeTransaction(
+		ctx,
+		s.datastore,
+		s.sdkClient,
+		s.kmsSigningClient,
+		s.kmsSigningKeyID,
+		fetchedTxn,
+	)
+	if err != nil {
+		return fmt.Errorf("failed to update transaction: %w", err)
+	}
+	stateMachine, err := StateMachineFromTransaction(s, authenticatedState)
+	if err != nil {
+		return fmt.Errorf("failed to create stateMachine: %w", err)
+	}
+	_, err = Drive(ctx, stateMachine)
+	if err != nil {
+		// Insufficient authorizations is an expected state. Treat it as such.
+		var insufficientAuthorizations *InsufficientAuthorizationsError
+		if errors.As(err, &insufficientAuthorizations) {
+			return nil
+		}
+		return fmt.Errorf("failed to progress transaction: %w", err)
+	}
+	// If the above call to Drive succeeds without giving insufficientAuthorizations,
+	// it's time to kick off payment. @TODO: Needs to be async, but for dry-run we
+	// can leave it synchronous.
+	_, err = Drive(ctx, stateMachine)
+	if err != nil {
+		return fmt.Errorf("failed to progress transaction: %w", err)
+	}
+
+	return nil
+}
