+++ services/payments/datastore.go
@@ -343,72 +343,6 @@ func (s Service) insertPayment(
 	return s.datastore.InsertPaymentState(ctx, paymentStateForSigning)
 }
 
-// AuthorizeTransaction - Add an Authorization for the Transaction and attempt to Drive
-// the Transaction forward. NOTE: This function assumes that the http signature has been
-// verified before running. This is achieved in the SubmitHandler middleware.
-func (s *Service) AuthorizeTransaction(
-	ctx context.Context,
-	keyID string,
-	transaction AuthenticatedPaymentState,
-) error {
-	fetchedTxn, idempotencyKey, err := s.GetTransactionFromDocumentID(ctx, transaction.DocumentID)
-	if err != nil {
-		return fmt.Errorf(
-			"failed to get transaction with idempotencyKey %s by document ID %s: %w",
-			idempotencyKey,
-			transaction.DocumentID,
-			err,
-		)
-	}
-	auth := PaymentAuthorization{
-		KeyID:      keyID,
-		DocumentID: transaction.DocumentID,
-	}
-	keyHasNotYetSigned := true
-	for _, authorization := range fetchedTxn.Authorizations {
-		if authorization.KeyID == auth.KeyID {
-			keyHasNotYetSigned = false
-		}
-	}
-	if !keyHasNotYetSigned {
-		return fmt.Errorf("key %s has already signed document %s", auth.KeyID, fetchedTxn.DocumentID)
-	}
-	fetchedTxn.Authorizations = append(fetchedTxn.Authorizations, auth)
-	authenticatedState, err := writeTransaction(
-		ctx,
-		s.datastore,
-		s.sdkClient,
-		s.kmsSigningClient,
-		s.kmsSigningKeyID,
-		fetchedTxn,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to update transaction: %w", err)
-	}
-	stateMachine, err := StateMachineFromTransaction(s, authenticatedState)
-	if err != nil {
-		return fmt.Errorf("failed to create stateMachine: %w", err)
-	}
-	_, err = Drive(ctx, stateMachine)
-	if err != nil {
-		// Insufficient authorizations is an expected state. Treat it as such.
-		var insufficientAuthorizations *InsufficientAuthorizationsError
-		if errors.As(err, &insufficientAuthorizations) {
-			return nil
-		}
-		return fmt.Errorf("failed to progress transaction: %w", err)
-	}
-	// If the above call to Drive succeeds without giving insufficientAuthorizations,
-	// it's time to kick off payment. @TODO: Needs to be async, but for dry-run we
-	// can leave it synchronous.
-	_, err = Drive(ctx, stateMachine)
-	if err != nil {
-		return fmt.Errorf("failed to progress transaction: %w", err)
-	}
-
-	return nil
-}
-
 // GetTransactionFromDocumentID - get the transaction data from the document ID in qldb.
 func (s *Service) GetTransactionFromDocumentID(
 	ctx context.Context,
