+++ services/payments/qldb_types.go
@@ -18,8 +18,7 @@ type QLDBPaymentTransitionHistoryEntryBlockAddress struct {
 // QLDBPaymentTransitionHistoryEntryHash defines hash for qldbPaymentTransitionHistoryEntry.
 type QLDBPaymentTransitionHistoryEntryHash []byte
 
-// qldbPaymentTransitionHistoryEntrySignature defines signature for
-// qldbPaymentTransitionHistoryEntry.
+// qldbPaymentTransitionHistoryEntrySignature defines signature for qldbPaymentTransitionHistoryEntry.
 type qldbPaymentTransitionHistoryEntrySignature []byte
 
 // QLDBPaymentTransitionHistoryEntryMetadata defines metadata for qldbPaymentTransitionHistoryEntry
